+++ src/ChowRNN/LayerRandomiser.hpp
@@ -4,6 +4,7 @@
 #include <MLUtils/Model.h>
 #include <random>
 
+// utility functions to randomise the weights of an RNN layer
 class LayerRandomiser {
 public:
     LayerRandomiser() = default;
@@ -22,7 +23,7 @@ private:
     std::uniform_real_distribution<float> denseWeightDist { -1.0f, 1.0f };
     std::uniform_real_distribution<float> biasDist { -0.5f, 0.5f };
     std::uniform_real_distribution<float> kernelDist { -1.0f, 1.0f };
-    std::uniform_real_distribution<float> recurrentDist { -0.9f, 0.9f };
+    std::uniform_real_distribution<float> recurrentDist { -0.5f, 0.5f };
 };
 
 #endif // LAYERRANDOMISER_H_INCLUDED