+++ CHANGELOG.md
@@ -3,6 +3,7 @@ All notable changes to this project will be documented in
 this file.
 
 ## [Unreleased]
+- New Recurrent Neural Network module.
 - Add 4x oversampling to Chow Tape.
 - Refactor delay lines to use Lagrange interpolation instead of sinc
   interpolation. This improves performance and removes dependency on